update to bench wheel_tree
